Factor the rest of CLI offline module out to clap-utils and cli-output crates
