+++ cli/tests/transfer.rs
@@ -1,10 +1,9 @@
 use solana_cli::{
     cli::{process_command, request_and_confirm_airdrop, CliCommand, CliConfig},
-    offline::parse_sign_only_reply_string,
     spend_utils::SpendAmount,
     test_utils::{check_ready, check_recent_balance},
 };
-use solana_cli_output::OutputFormat;
+use solana_cli_output::{parse_sign_only_reply_string, OutputFormat};
 use solana_client::{
     blockhash_query::{self, BlockhashQuery},
     nonce_utils,
